before javascript update

+++ app/controllers/Comments.java
@@ -1,9 +1,12 @@
 package controllers;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import models.Activity;
 import models.Comment;
+import models.CommentReply;
 import models.Event;
 import models.FileUpload;
 import models.Listing;
@@ -11,12 +14,91 @@ import models.User;
 
 import org.apache.commons.lang.StringEscapeUtils;
 
-import play.mvc.With;
+import utils.JsonUtils;
 import utils.RandomUtil;
+import utils.StringUtils;
 
-@With(Secure.class)
+import com.google.gson.JsonObject;
+
+import dto.CommentDTO;
+
 public class Comments extends BaseController
 {
+    public static void getComments(String type, String listing, String event)
+    {
+        final User user = getLoggedUser();
+        final Integer first = request.params.get("first") != null ? Integer.parseInt(request.params.get("first")) : 0;
+        final Integer count = request.params.get("count") != null ? Integer.parseInt(request.params.get("count")) : 4;
+
+        List<Comment> comments = null;
+        if (type.equals("event"))
+        {
+            Event e = Event.get(event);
+            comments = Comment.getByEvent(e, first, count);
+        }
+        else if (type.equals("listing"))
+        {
+            Listing l = Listing.get(listing);
+            comments = Comment.getByListing(l, first, count);
+        }
+        else if (type.equals("user") && user != null)
+        {
+            comments = Comment.getByFollower(user, first, count);
+        }
+
+        List<CommentDTO> commentsDto = new ArrayList<CommentDTO>();
+        for (Comment comment : comments)
+            commentsDto.add(CommentDTO.convert(comment, user));
+
+        renderJSON(commentsDto);
+    }
+
+    public static void addReply()
+    {
+        final JsonObject body = JsonUtils.getJson(request.body);
+        final User user = getLoggedUser();
+        final String id = body.get("id").getAsString();
+        final String comment = body.get("comment").getAsString();
+
+        if (user != null)
+        {
+            Comment c = Comment.getByUuid(id);
+            CommentReply cr = new CommentReply();
+            cr.comment = StringUtils.htmlEscape(comment);
+            cr.user = user;
+            cr.created = new Date();
+            cr.uuid = RandomUtil.getUUID();
+            cr.save();
+
+            if (c.event != null)
+            {
+                final Activity act = new Activity();
+                act.type = Activity.ACTIVITY_EVENT_COMMENTED;
+                act.user = user;
+                act.event = c.event;
+                act.eventName = c.event.listing.title;
+                act.saveActivity();
+            }
+
+            c.replies.add(cr);
+            c.updated = new Date();
+            c.save();
+        }
+        renderJSON("{\"response\":\"ok\"}");
+    }
+
+    public static void deleteReply(String uuid)
+    {
+        final User user = getLoggedUser();
+        if (user != null)
+        {
+            CommentReply cr = CommentReply.get(uuid);
+            if (cr != null)
+                cr.delete();
+        }
+        renderJSON("{\"response\":\"ok\"}");
+    }
+
     public static void addComment(String uuid, String objectType, String comment, String type, String url, String tempId)
     {
         final User user = getLoggedUser();
@@ -67,20 +149,17 @@ public class Comments extends BaseController
 
     public static void deleteComment(String uuid, String url)
     {
-        try
+        final User user = getLoggedUser();
+        Comment c = Comment.getByUuid(uuid);
+        if (c.canDelete(user))
         {
-            Comment c = Comment.getByUuid(uuid);
             List<FileUpload> files = c.files;
             for (FileUpload fileUpload : files)
             {
                 FileUpload.deleteOnDisc(fileUpload);
-                // fileUpload.delete();
             }
             c.delete();
-            redirectTo(url);
-        } catch (Exception e)
-        {
-            redirectTo(url);
         }
+        renderText("ok");
     }
 }
